backend/requirements.txt
Django==1.8
django-cors-headers==2.1.0
djangorestframework==3.6.3
